fix uint div int

+++ code/Src/motor.h
@@ -42,12 +42,12 @@ inline static void motor_offset(struct motor *s, int32_t offset_mm)
 
 inline static int16_t motor_get_pos(struct motor *s)
 {
-    return s->pos_step / s->k_mov2step;
+    return s->pos_step / (int16_t)s->k_mov2step;
 }
 
 inline static int16_t motor_get_offset(struct motor *s)
 {
-    return s->offset_step / s->k_mov2step;
+    return s->offset_step / (int16_t)s->k_mov2step;
 }
 
 inline static void motor_set_velocity(struct motor *s, uint32_t vel_papugay)
